requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
google-generativeai==0.8.2
google-genai>=1.0.0